CamelCase renaming and and warnings-fixes (#42)
