+++ client/src/components/Banner.tsx
