+++ client/src/components/MenuMonths.tsx
@@ -1,18 +1,15 @@
 import React from 'react';
 
-
-interface IMenuMonths {
-    links: string[12]
-}
 const monthNames = [ "January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December" ];
 
 function MenuMonths() {
-    return (
+  return (
       <ul style={{listStyleType: "none"}}>
+        {/* eslint-disable jsx-a11y/anchor-is-valid */}
         <li><a href="#">Senaste nytt</a></li>
         {monthNames.map(month => <li><a href='#'>{month}</a></li>)}
       </ul>
     )
   }
-  
-  export default MenuMonths;
+
+export default MenuMonths;
