+++ client/src/components/news/NewsFeedTiny.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import NewsItemTiny, {INewsItemTiny} from './NewsItemTiny';
-import SomeButton, {IButton} from '../button';
+import SomeButton, {IButton} from '../Button';
 import Container from 'react-bootstrap/Container';
 
 
@@ -14,10 +14,16 @@ const DummyData : IButton = {type: "secondary", text: "Fler nyheter"}
 
 
 function NewsFeedTiny(props : INewsFeedTiny) {
+  const tinyFeed = props.items.map((item, index) =>
+    <li key={index} className="mb-3" style={{backgroundColor: "#f0f0f0"}}>
+      <NewsItemTiny {...item}/>
+    </li>
+  )
+
   return (
     <Container>
       <ul className="list-unstyled">
-        { props.items.map(item => <NewsItemTiny {...item}/>) }
+        {tinyFeed}
       </ul>
       <div className="text-center">
         <SomeButton {...DummyData}/>
